requirements.txt
# Core AI Framework
langgraph==0.2.28
langchain==0.2.16
langchain-openai==0.1.23
langchain-community==0.2.16

# API & Backend
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
websockets==13.1

# Database
sqlalchemy==2.0.35
alembic==1.13.3
aiosqlite==0.20.0

# Search & External APIs
tavily-python==0.5.0
openai==1.51.2

# Frontend
streamlit==1.39.0

# Utilities
python-dotenv==1.0.1
httpx==0.27.2
aiofiles==24.1.0
markdown==3.7
beautifulsoup4==4.12.3
lxml==5.3.0

# Development & Testing
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.7.0

# Additional utilities
python-multipart==0.0.12

